upgrade create-release ubuntu versions to 22.04
There seems to be problems in gnupg client talking to the gpg-agent when there's a significant
version mismatch between host and guest. Tested with 2.2.27 :: 2.2.34.

Note that this changes the docker image which means that the host key used by sshd will change and
you will get a nastygram when launching ssh about host identification changing. This is
expected. you should remove the previous host key.